requirements.txt
python-dotenv==1.0.1
tidalapi==0.7.3
deezer-python==6.1.1
